chore: Move XML utils into separate folder/namespace
