+++ lib/src/index.ts
@@ -12,3 +12,4 @@ export * from './id3.js';
 export type * from './request.js';
 export * from './structuredfield.js';
 export * from './utils.js';
+export * from './xml.js';
