+++ lib/src/utils.ts
@@ -5,7 +5,6 @@
  */
 export * from './utils/base64decode.js';
 export * from './utils/base64encode.js';
-export * from './utils/parseXml.js';
 export * from './utils/roundToEven.js';
 export * from './utils/unescapeHtml.js';
 export * from './utils/urlToRelativePath.js';
