+++ lib/src/xml.ts
@@ -1,0 +1,12 @@
+/**
+ * A collection of tools for working with XML.
+ *
+ * @groupDescription XML
+ * @packageDocumentation
+ * @beta
+ */
+export * from './xml/parseXml.js';
+export * from './xml/XmlChildren.js';
+export * from './xml/XmlNode.js';
+export * from './xml/XmlParseOptions.js';
+
