+++ lib/src/xml/XmlChildren.ts
@@ -1,0 +1,10 @@
+import type { XmlNode } from './XmlNode';
+
+/**
+ * XML children list
+ *
+ * @group XML
+ *
+ * @beta
+ */
+export type XmlChildren = (XmlNode | string)[];
