+++ lib/src/xml/XmlNode.ts
@@ -1,0 +1,14 @@
+import type { XmlChildren } from './XmlChildren';
+
+/**
+ * XML node
+ *
+ * @group XML
+ *
+ * @beta
+ */
+export type XmlNode = {
+	tagName: string;
+	attributes: Record<string, string>;
+	children: XmlChildren;
+};
