+++ lib/src/xml/XmlParseOptions.ts
@@ -1,0 +1,12 @@
+/**
+ * XML parsing options
+ *
+ * @group XML
+ *
+ * @beta
+ */
+export type XmlParseOptions = {
+	pos?: number;
+	keepWhitespace?: boolean;
+	keepComments?: boolean;
+};
