+++ lib/test/xml/fixtures/node_types.xml
